# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "مستندات API سامانه لد ",
        "description": "توضیحات نحوه استفاده از APIها",
        "contact": {
            "email": "hreza779@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "سرور اصلی API"
        }
    ],
    "paths": {
...
        "/api/exam-assignments": {
            "get": {
                "tags": [
                    "Exam Assignments"
                ],
                "summary": "Get all exam assignments",
                "operationId": "282335bc3e829ff51884dc79c5c193cf",
                "parameters": [
                    {
                        "name": "exam_id",
                        "in": "query",
                        "description": "Filter by exam ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "manager_id",
                        "in": "query",
                        "description": "Filter by manager ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "assigned",
                                "started",
                                "completed",
                                "expired"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of exam assignments"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Exam Assignments"
                ],
                "summary": "Assign exam to manager(s)",
                "operationId": "031983dc0be420569e862153df6f1d61",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "exam_id",
                                    "manager_ids"
                                ],
                                "properties": {
                                    "exam_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "manager_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-02-01"
                                    },
                                    "max_attempts": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exam assigned successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exam-assignments/{assignment}": {
            "get": {
                "tags": [
                    "Exam Assignments"
                ],
                "summary": "Get exam assignment details",
                "operationId": "51c7cf547b74f11635c9a904dda38d2e",
                "parameters": [
                    {
                        "name": "assignment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assignment details"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Exam Assignments"
                ],
                "summary": "Update exam assignment",
                "operationId": "f1096a593cacc129bd5df81a42ebd781",
                "parameters": [
                    {
                        "name": "assignment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "max_attempts": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "assigned",
                                            "started",
                                            "completed",
                                            "expired"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assignment updated successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Exam Assignments"
                ],
                "summary": "Delete exam assignment",
                "operationId": "457a5213a015ae06035a2b5b74f5f817",
                "parameters": [
                    {
                        "name": "assignment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Assignment deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exam-assignments/{assignment}/start": {
            "post": {
                "tags": [
                    "Exam Assignments"
                ],
                "summary": "Mark exam as started",
                "operationId": "154de778a56d91533ff9c84d505e79b4",
                "parameters": [
                    {
                        "name": "assignment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam started"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exam-assignments/{assignment}/complete": {
            "post": {
                "tags": [
                    "Exam Assignments"
                ],
                "summary": "Mark exam as completed",
                "operationId": "28e0eecadb210594bc4ec51955075833",
                "parameters": [
                    {
                        "name": "assignment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam completed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}